+++ lib/facebook/messenger.rb
@@ -1,9 +1,10 @@
 require 'facebook/messenger/version'
 require 'facebook/messenger/error'
 require 'facebook/messenger/subscriptions'
-require 'facebook/messenger/thread'
+require 'facebook/messenger/profile'
 require 'facebook/messenger/bot'
 require 'facebook/messenger/server'
+require 'facebook/messenger/server_no_error'
 require 'facebook/messenger/configuration'
 require 'facebook/messenger/incoming'
 
@@ -22,6 +23,8 @@ module Facebook
       @config = config
     end
 
+    # Set the default configuration provider.
+    # Developer can set different configuration provider.
     configure do |config|
       config.provider = Configuration::Providers::Environment.new
     end