+++ lib/facebook/messenger/bot.rb
@@ -1,20 +1,44 @@
+require 'facebook/messenger/bot/error_parser'
+require 'facebook/messenger/bot/exceptions'
+require 'facebook/messenger/bot/message_type'
+require 'facebook/messenger/bot/messaging_type'
+require 'facebook/messenger/bot/tag'
+
 module Facebook
   module Messenger
-    # The Bot module sends and receives messages.
+    #
+    # Module Bot provides functionality to sends and receives messages.
+    #
     module Bot
       include HTTParty
 
-      base_uri 'https://graph.facebook.com/v2.6/me'
+      # Define base_uri for HTTParty.
+      base_uri 'https://graph.facebook.com/v2.9/me'
 
-      EVENTS = [:message, :delivery, :postback, :optin,
-                :read, :account_linking].freeze
+      #
+      # @return [Array] Array containing the supported webhook events.
+      EVENTS = %i[
+        message
+        delivery
+        postback
+        optin
+        read
+        account_linking
+        referral
+        message_echo
+        payment
+        policy_enforcement
+      ].freeze
 
       class << self
         # Deliver a message with the given payload.
+        # @see https://developers.facebook.com/docs/messenger-platform/send-api-reference#request
         #
-        # message - A Hash describing the recipient and the message*.
+        # @raise [Facebook::Messenger::Bot::SendError] if there is any error
+        #   in response while sending message.
         #
-        # * https://developers.facebook.com/docs/messenger-platform/send-api-reference#request
+        # @param [Hash] message A Hash describing the recipient and the message.
+        # @param [String] access_token Access token.
         #
         # Returns a String describing the message ID if the message was sent,
         # or raises an exception if it was not.
@@ -26,15 +50,19 @@ module Facebook
                             access_token: access_token
                           }
 
-          raise_errors_from(response)
+          Facebook::Messenger::Bot::ErrorParser.raise_errors_from(response)
 
-          response['message_id']
+          response.body
         end
 
         # Register a hook for the given event.
         #
-        # event - A String describing a Messenger event.
-        # block - A code block to run upon the event.
+        # @raise [ArgumentError] if received event is not registered.
+        #
+        # @param [String] event A String describing a Messenger event.
+        # @param [Block] block A code block to run upon the event.
+        #
+        # @return Save event and its block in hooks.
         def on(event, &block)
           unless EVENTS.include? event
             raise ArgumentError,
@@ -47,9 +75,12 @@ module Facebook
 
         # Receive a given message from Messenger.
         #
-        # payload - A Hash describing the message.
+        # @see https://developers.facebook.com/docs/messenger-platform/webhook-reference
         #
-        # * https://developers.facebook.com/docs/messenger-platform/webhook-reference
+        # @param [Hash] payload A Hash describing the message.
+        #
+        # @return pass event and object of callback class to trigger function.
+        #
         def receive(payload)
           callback = Facebook::Messenger::Incoming.parse(payload)
           event = Facebook::Messenger::Incoming::EVENTS.invert[callback.class]
@@ -57,66 +88,50 @@ module Facebook
         end
 
         # Trigger the hook for the given event.
+        # Fetch callback for event from hooks and call it.
         #
-        # event - A String describing a Messenger event.
-        # args - Arguments to pass to the hook.
+        # @raise [KeyError] if hook is not registered for event
+        #
+        # @param [String] event A String describing a Messenger event.
+        # @param [Object] args Arguments to pass to the hook.
         def trigger(event, *args)
           hooks.fetch(event).call(*args)
         rescue KeyError
-          $stderr.puts "Ignoring #{event} (no hook registered)"
+          warn "Ignoring #{event} (no hook registered)"
         end
 
-        # Raise any errors in the given response.
         #
-        # response - A HTTParty::Response object.
-        #
-        # Returns nil if no errors were found, otherwises raises appropriately.
-        def raise_errors_from(response)
-          return unless response.key? 'error'
-          error = response['error']
-
-          raise(
-            error_class_from_error_code(error['code']),
-            (error['error_data'] || error['message'])
-          )
-        end
-
-        # Find the appropriate error class for the given error code.
-        #
-        # error_code - An Integer describing an error code.
-        #
-        # Returns an error class, or raises KeyError if none was found.
-        def error_class_from_error_code(error_code)
-          {
-            100 => RecipientNotFound,
-            10 => PermissionDenied,
-            2 => InternalError
-          }[error_code] || Facebook::Messenger::Error
-        end
-
         # Return a Hash of hooks.
+        #
+        # @return [Hash] Hash of hooks.
+        #
         def hooks
           @hooks ||= {}
         end
 
+        #
         # Deregister all hooks.
+        #
+        # @return [Hash] Assign empty hash to hooks and return it.
+        #
         def unhook
           @hooks = {}
         end
 
+        #
         # Default HTTParty options.
+        #
+        # @return [Hash] Default HTTParty options.
+        #
         def default_options
           super.merge(
+            read_timeout: 300,
             headers: {
               'Content-Type' => 'application/json'
             }
           )
         end
       end
-
-      class RecipientNotFound < Facebook::Messenger::Error; end
-      class PermissionDenied < Facebook::Messenger::Error; end
-      class InternalError < Facebook::Messenger::Error; end
     end
   end
 end